pyproject.toml
[project]
name = "draft-mail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "chromadb>=1.3.5",
    "deepagents>=0.3.0",
    "docx>=0.2.4",
    "fastapi>=0.124.4",
    "gradio>=6.1.0",
    "jose>=1.0.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.15",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.10",
    "langsmith>=0.4.45",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pinecone>=7.3.0",
    "pydantic[email]>=2.12.4",
    "pypdf>=6.4.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.51.0",
    "tavily>=1.1.0",
    "uvicorn>=0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["draft-mail"]
